Tasks and faculties_page.dart errors

+++ final_project/lib/services/sqllite.dart
@@ -1,5 +1,5 @@
-import 'package:sqflite/sqflite.dart' as sql;
 import 'package:path/path.dart' as path;
+import 'package:sqflite/sqflite.dart' as sql;
 
 import '../constants.dart';
 import '../model/task.dart';
@@ -7,10 +7,8 @@ import '../model/task.dart';
 class Sqlite {
   static Future<sql.Database> _initDatabase() async {
     final sqlPath = await sql.getDatabasesPath();
-    return sql.openDatabase(path.join(sqlPath, 'localDatabase.db'),
-        onCreate: (db, version) {
-      return db
-          .execute('''CREATE TABLE tasks (id INTEGER PRIMARY KEY AUTOINCREMENT,
+    return sql.openDatabase(path.join(sqlPath, 'localDatabase.db'), onCreate: (db, version) {
+      return db.execute('''CREATE TABLE tasks (id INTEGER PRIMARY KEY AUTOINCREMENT,
           uid TEXT NOT NULL, 
         body TEXT NOT NULL,
         reminderIsSet TEXT NOT NULL, 
@@ -51,8 +49,7 @@ class Sqlite {
 
   Future<int> delete(int id) async {
     final db = await Sqlite._initDatabase();
-    return await db.delete('tasks',
-        where: 'id = ? AND uid = ?', whereArgs: [id, Constants.uid]);
+    return await db.delete('tasks', where: 'id = ? AND uid = ?', whereArgs: [id, Constants.uid]);
   }
 
   Future<int> update(Task task) async {
